move lora sa to models dir, enforce non-empty lora_modules arg
