+++ docs/source/api_ref_models.rst
@@ -6,6 +6,7 @@ torchtune.models
 
 .. autosummary::
     :toctree: generated/
-    :template: class.rst
+    :template: functions.rst
 
     llama2_7b
+    lora_llama_self_attention
