+++ torchtune/modules/peft/__init__.py
@@ -4,11 +4,10 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
-from .lora import FusedLoRADim, lora_llama_self_attention, LoRAFusedLinear, LoRALinear
+from .lora import FusedLoRADim, LoRAFusedLinear, LoRALinear
 
 __all__ = [
     "FusedLoRADim",
     "LoRAFusedLinear",
     "LoRALinear",
-    "lora_llama_self_attention",
 ]
